chore: make SegmentedControl reusable
